Ajout du fichier tri.py et implementation dans stat_1.py

+++ affichage.py
@@ -6,7 +6,7 @@ noteVal = st.Statistique()
 
 def valuer(): 
     try:
-        noteVal.ajout(int(valeurText.get()), int(effectifText.get()))
+        noteVal.ajout(float(valeurText.get()), int(effectifText.get()))
     except:
         messagebox.showerror("Erreur", "Veiller mettre des informations correct !")
         return 0
